added address deletion

+++ CustomerManagement/Services/AddressService.cs
@@ -26,6 +26,24 @@ namespace CustomerManagement.Services
 
         public void Delete(int customerId, int addressId)
         {
+            IEnumerable<Address> addresses = GetAllByCustomerId(customerId);
+            if(addresses.Count() <= Constants.ADDRESS_MINIMUM_AMOUNT)
+            {
+                logger.LogError("The customer has only one address. The min number of addresses is {}. CustomerId is {}, addressId is {}",
+                     Constants.ADDRESS_MINIMUM_AMOUNT, customerId, addressId);
+                //throw new Exception();
+                return;
+            }
+
+            Address dbAddress = GetById(customerId, addressId);
+            if (dbAddress.IsMain)
+            {
+                logger.LogError("Main address could not be removed. CustomerId is {}, addressId is {}",
+                     customerId, addressId);
+                //throw new Exception();
+                return;
+            }
+
             addressRepository.Delete(customerId, addressId);
         }
 
@@ -41,7 +59,15 @@ namespace CustomerManagement.Services
 
         public Address GetMainAddress(int customerId)
         {
-            throw new NotImplementedException();
+            IEnumerable<Address> addresses = addressRepository.GetAllByCustomerId(customerId);
+            if(addresses.Count() == 1)
+            {
+                return addresses.First();
+            }
+            else
+            {
+                return addresses.Where(address => address.IsMain).First();
+            }
         }
 
         public Address Update(int customerId, Address address)
